refactor: centralize backend serialization logic
Extract all logic for converting frontend state into the JSON format expected by the C++ backend into a new `serializers.ts` module.

This removes significant code duplication from `backendSlice` (for full scenario serialization) and `scenarioSlice` (for granular item updates). The granular update logic is also improved to correctly handle individual platform components and `Timing` objects.
